+++ e2e/packages/contracts/types/ethers-contracts/factories/IWorld__factory.ts
@@ -26,6 +26,22 @@ const _abi = [
   {
     inputs: [
       {
+        internalType: "address",
+        name: "delegator",
+        type: "address",
+      },
+      {
+        internalType: "address",
+        name: "delegatee",
+        type: "address",
+      },
+    ],
+    name: "DelegationNotFound",
+    type: "error",
+  },
+  {
+    inputs: [
+      {
         internalType: "bytes4",
         name: "functionSelector",
         type: "bytes4",
@@ -369,6 +385,35 @@ const _abi = [
       },
     ],
     name: "call",
+    outputs: [
+      {
+        internalType: "bytes",
+        name: "",
+        type: "bytes",
+      },
+    ],
+    stateMutability: "payable",
+    type: "function",
+  },
+  {
+    inputs: [
+      {
+        internalType: "address",
+        name: "delegator",
+        type: "address",
+      },
+      {
+        internalType: "bytes32",
+        name: "resourceSelector",
+        type: "bytes32",
+      },
+      {
+        internalType: "bytes",
+        name: "funcSelectorAndArgs",
+        type: "bytes",
+      },
+    ],
+    name: "callFrom",
     outputs: [
       {
         internalType: "bytes",
@@ -770,6 +815,29 @@ const _abi = [
   {
     inputs: [
       {
+        internalType: "address",
+        name: "delegatee",
+        type: "address",
+      },
+      {
+        internalType: "bytes32",
+        name: "delegationControlId",
+        type: "bytes32",
+      },
+      {
+        internalType: "bytes",
+        name: "initFuncSelectorAndArgs",
+        type: "bytes",
+      },
+    ],
+    name: "registerDelegation",
+    outputs: [],
+    stateMutability: "nonpayable",
+    type: "function",
+  },
+  {
+    inputs: [
+      {
         internalType: "bytes32",
         name: "resourceSelector",
         type: "bytes32",
@@ -864,7 +932,7 @@ const _abi = [
         type: "bytes32",
       },
       {
-        internalType: "contract System",
+        internalType: "contract WorldContextConsumer",
         name: "system",
         type: "address",
       },
@@ -1039,6 +1107,24 @@ const _abi = [
       },
     ],
     stateMutability: "pure",
+    type: "function",
+  },
+  {
+    inputs: [
+      {
+        internalType: "bytes16",
+        name: "namespace",
+        type: "bytes16",
+      },
+      {
+        internalType: "address",
+        name: "newOwner",
+        type: "address",
+      },
+    ],
+    name: "transferOwnership",
+    outputs: [],
+    stateMutability: "nonpayable",
     type: "function",
   },
   {