+++ e2e/packages/sync-test/indexerSync.test.ts
@@ -23,6 +23,17 @@ import {
 import { range } from "@latticexyz/utils";
 import path from "node:path";
 import { rpcHttpUrl } from "./setup/constants";
+import { z } from "zod";
+
+const env = z
+  .object({
+    DATABASE_URL: z.string().default("postgres://127.0.0.1/postgres"),
+  })
+  .parse(process.env, {
+    errorMap: (issue) => ({
+      message: `Missing or invalid environment variable: ${issue.path.join(".")}`,
+    }),
+  });
 
 describe("Sync from indexer", async () => {
   const asyncErrorHandler = createAsyncErrorHandler();
@@ -54,14 +65,21 @@ describe("Sync from indexer", async () => {
     expect(asyncErrorHandler.getErrors()[0]).toContain("error fetching initial state from indexer");
   });
 
-  describe("indexer online", () => {
+  describe.each([["sqlite"], ["postgres"]] as const)("%s indexer", (indexerType) => {
     let indexerIteration = 1;
     let indexer: ReturnType<typeof startIndexer>;
 
     beforeEach(async () => {
       // Start indexer
       const port = 3000 + indexerIteration++;
-      indexer = startIndexer(port, path.join(__dirname, `anvil-${port}.db`), rpcHttpUrl, asyncErrorHandler.reportError);
+      indexer = startIndexer({
+        port,
+        rpcHttpUrl,
+        reportError: asyncErrorHandler.reportError,
+        ...(indexerType === "postgres"
+          ? { indexer: "postgres", databaseUrl: env.DATABASE_URL }
+          : { indexer: "sqlite", sqliteFilename: path.join(__dirname, `anvil-${port}.db`) }),
+      });
       await indexer.doneSyncing;
     });