+++ packages/world/src/modules/core/implementations/EphemeralRecordSystem.sol
@@ -3,13 +3,9 @@ pragma solidity >=0.8.0;
 
 import { IStoreEphemeral } from "@latticexyz/store/src/IStore.sol";
 import { Schema } from "@latticexyz/store/src/Schema.sol";
-import { IModule } from "../../../interfaces/IModule.sol";
 import { System } from "../../../System.sol";
 import { ResourceSelector } from "../../../ResourceSelector.sol";
 import { AccessControl } from "../../../AccessControl.sol";
-import { Call } from "../../../Call.sol";
-import { ResourceAccess } from "../../../tables/ResourceAccess.sol";
-import { InstalledModules } from "../../../tables/InstalledModules.sol";
 import { StoreCore } from "@latticexyz/store/src/StoreCore.sol";
 
 contract EphemeralRecordSystem is IStoreEphemeral, System {